requirements.txt
fastapi==0.104.1
line-bot-sdk==3.16.3
httpx==0.25.2
python-dotenv==1.0.0
